requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy>=2.0
asyncpg
python-dotenv